requirements.txt
Package                   Version
------------------------- ---------
altgraph                  0.17.4
annotated-types           0.7.0
anyio                     4.8.0
asyncio                   3.4.3
certifi                   2024.8.30
charset-normalizer        3.4.0
colorama                  0.4.6
distro                    1.9.0
h11                       0.14.0
httpcore                  1.0.7
httpx                     0.28.1
idna                      3.10
jiter                     0.8.2
mypy                      1.14.0
mypy-extensions           1.0.0
openai                    1.59.6
packaging                 24.2
pefile                    2023.2.7
pillow                    11.1.0
pillow_heif               0.21.0
pip                       24.3.1
plyer                     2.1.0
pydantic                  2.10.5
pydantic_core             2.27.2
pyinstaller               6.11.1
pyinstaller-hooks-contrib 2024.10
pywin32-ctypes            0.2.3
PyYAML                    6.0.2
requests                  2.32.3
setuptools                75.6.0
sniffio                   1.3.1
tqdm                      4.67.1
typing_extensions         4.12.2
urllib3                   2.2.3
